Flatten zip operator output explicitly
The output of the zip operator `Z` used to be flattened implicitly.  This is
no longer the case, thus to reinstate the previous behaviour expected in the
context tests, the zipped lists need to be explicitly flattened before their
values are passed into the relevant for loop.  This change makes the
`context.t` tests pass.